back-end/requirements.txt
tensorflow==2.15.0 
flask
pyngrok
flask-cors
gunicorn
Flask-Limiter
google-cloud-firestore